Small tweaks on OutsideLevel. Added checkpoints.
Fixed a graphical glitch on the gateway into the forge.
Changed the size of the bridge torchiers.
Changed the size of the FireStatues.
Added a floor under the lava for the player to land on and die.
Added some walls to make things look less buggy while under the lava.
Still doesn't look good but is much better.